menu: module refactoring
split menu into static and dynamic part
put both parts into separate files
command arrays are put into c files
command registration via register utility functions
ua_event_handler and related funtions stay in menu.c
