+++ modules/menu/module.mk
@@ -5,6 +5,6 @@
 #
 
 MOD		:= menu
-$(MOD)_SRCS	+= menu.c
+$(MOD)_SRCS	+= menu.c static_menu.c dynamic_menu.c
 
 include mk/mod.mk
